Restructuring statistic collectors
