+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/CallbackHandledEvent.java
@@ -19,7 +19,7 @@
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
 import org.apache.synapse.MessageContext;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.CallbackStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 
@@ -38,6 +38,6 @@ public class CallbackHandledEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		RuntimeStatisticCollector.removeCallback(statisticId, callbackId);
+		CallbackStatisticCollector.removeCallback(statisticId, callbackId);
 	}
 }
