+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/CallbackSentEvent.java
@@ -19,7 +19,7 @@
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
 import org.apache.synapse.MessageContext;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.CallbackStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 
@@ -44,6 +44,6 @@ public class CallbackSentEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		RuntimeStatisticCollector.addCallbacks(statisticId, callbackId, cloneId);
+		CallbackStatisticCollector.addCallbacks(statisticId, callbackId, cloneId);
 	}
 }
