+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/EndFlowEvent.java
@@ -38,6 +38,6 @@ public class EndFlowEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		RuntimeStatisticCollector.closeStatisticEntryForcefully(statisticDataUnit);
+		RuntimeStatisticCollector.closeStatisticEntry(statisticDataUnit, StatisticsConstants.FORECEFULLY_CLOSE);
 	}
 }
