+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/EndpointEvent.java
@@ -19,7 +19,7 @@
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
 import org.apache.synapse.MessageContext;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.EndpointStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 import org.apache.synapse.core.SynapseEnvironment;
@@ -49,7 +49,7 @@ public class EndpointEvent implements StatisticsReportingEvent {
 	}
 
 	@Override public void process() {
-		RuntimeStatisticCollector
+		EndpointStatisticCollector
 				.createEndpointStatistics(statisticId, timeStamp, endpointId, endpointName, synapseEnvironment, time,
 				                          isCreateLog);
 	}
