+++ modules/core/src/main/java/org/apache/synapse/core/axis2/SynapseMessageReceiver.java
@@ -29,7 +29,7 @@ import org.apache.synapse.SynapseConstants;
 import org.apache.synapse.SynapseException;
 import org.apache.synapse.aspects.AspectConfigurationDetectionStrategy;
 import org.apache.synapse.aspects.ComponentType;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.ProxyStatisticCollector;
 import org.apache.synapse.aspects.statistics.StatisticsReporter;
 
 /**
@@ -94,7 +94,7 @@ public class SynapseMessageReceiver implements MessageReceiver {
             }
         } finally {
             StatisticsReporter.endReportForAllOnRequestProcessed(synCtx);
-            RuntimeStatisticCollector.reportEndSynapseMessageReceiver(synCtx);
+            ProxyStatisticCollector.reportEndSynapseMessageReceiver(synCtx);
         }
     }
