+++ modules/core/src/main/java/org/apache/synapse/core/axis2/TimeoutHandler.java
@@ -27,7 +27,7 @@ import org.apache.synapse.FaultHandler;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.ServerContextInformation;
 import org.apache.synapse.SynapseConstants;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.CallbackStatisticCollector;
 import org.apache.synapse.aspects.statistics.StatisticsCleaner;
 import org.apache.synapse.aspects.statistics.StatisticsCollector;
 import org.apache.synapse.config.SynapseConfigUtils;
@@ -211,7 +211,7 @@ public class TimeoutHandler extends TimerTask {
                                 "timeout of : " + (callback.getTimeoutDuration() / 1000) + " seconds");
                     }
                     callbackStore.remove(key);
-                    RuntimeStatisticCollector.reportCallbackReceived(callback.getSynapseOutMsgCtx(), (String) key);
+                    CallbackStatisticCollector.reportCallbackReceived(callback.getSynapseOutMsgCtx(), (String) key);
                 }
             }
         }
