+++ modules/core/src/main/java/org/apache/synapse/mediators/builtin/CacheMediator.java
@@ -48,7 +48,7 @@ import org.apache.synapse.Mediator;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.SynapseException;
 import org.apache.synapse.SynapseLog;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.MediatorStatisticCollector;
 import org.apache.synapse.continuation.ContinuationStackManager;
 import org.apache.synapse.core.SynapseEnvironment;
 import org.apache.synapse.core.axis2.Axis2MessageContext;
@@ -202,14 +202,14 @@ public class CacheMediator extends AbstractMediator implements ManagedLifecycle,
         if (synLog.isTraceOrDebugEnabled()) {
             synLog.traceOrDebug("Aggregate mediator : Mediating from ContinuationState");
         }
-        RuntimeStatisticCollector
+        MediatorStatisticCollector
                 .openLogForContinuation(synCtx, onCacheHitSequence.getSequenceNameForStatistics(synCtx));
         if (!contState.hasChild()) {
             onCacheHitSequence.mediate(synCtx, contState.getPosition() + 1);
         } else {
             FlowContinuableMediator mediator = (FlowContinuableMediator) onCacheHitSequence.
                     getChild(contState.getPosition());
-            RuntimeStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
+            MediatorStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
 
             mediator.mediate(synCtx, contState.getChildContState());
