+++ modules/core/src/main/java/org/apache/synapse/mediators/filters/SwitchMediator.java
@@ -21,7 +21,7 @@ package org.apache.synapse.mediators.filters;
 
 import org.apache.synapse.ContinuationState;
 import org.apache.synapse.Mediator;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.MediatorStatisticCollector;
 import org.apache.synapse.config.xml.SynapsePath;
 import org.apache.synapse.continuation.ContinuationStackManager;
 import org.apache.synapse.continuation.ReliantContinuationState;
@@ -176,7 +176,7 @@ public class SwitchMediator extends AbstractMediator implements ManagedLifecycle
                 FlowContinuableMediator mediator =
                         (FlowContinuableMediator) defaultCase.getCaseMediator().
                                 getChild(continuationState.getPosition());
-                RuntimeStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
+                MediatorStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
 
                 result = mediator.mediate(synCtx, continuationState.getChildContState());
 
@@ -190,7 +190,7 @@ public class SwitchMediator extends AbstractMediator implements ManagedLifecycle
                 FlowContinuableMediator mediator =
                         (FlowContinuableMediator) cases.get(subBranch - 1).getCaseMediator().
                                 getChild(continuationState.getPosition());
-                RuntimeStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
+                MediatorStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
 
                 result = mediator.mediate(synCtx, continuationState.getChildContState());
