+++ modules/core/src/main/java/org/apache/synapse/mediators/template/InvokeMediator.java
@@ -23,7 +23,7 @@ import org.apache.synapse.ManagedLifecycle;
 import org.apache.synapse.Mediator;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.SynapseLog;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.MediatorStatisticCollector;
 import org.apache.synapse.continuation.ContinuationStackManager;
 import org.apache.synapse.continuation.ReliantContinuationState;
 import org.apache.synapse.core.SynapseEnvironment;
@@ -156,7 +156,7 @@ public class InvokeMediator extends AbstractMediator implements
         if (subBranch == 0) {
 	        // Default flow
 	        TemplateMediator templateMediator = (TemplateMediator) synCtx.getSequenceTemplate(targetTemplate);
-	        RuntimeStatisticCollector.openLogForContinuation(synCtx, templateMediator.getMediatorName());
+	        MediatorStatisticCollector.openLogForContinuation(synCtx, templateMediator.getMediatorName());
 	        if (!continuationState.hasChild()) {
 		        result = templateMediator.mediate(synCtx, continuationState.getPosition() + 1);
 		        if (result) {
@@ -165,7 +165,7 @@ public class InvokeMediator extends AbstractMediator implements
 	        } else {
 		        FlowContinuableMediator mediator =
 				        (FlowContinuableMediator) templateMediator.getChild(continuationState.getPosition());
-		        RuntimeStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
+		        MediatorStatisticCollector.openLogForContinuation(synCtx, ((Mediator) mediator).getMediatorName());
 
 		        result = mediator.mediate(synCtx, continuationState.getChildContState());
 
@@ -176,7 +176,7 @@ public class InvokeMediator extends AbstractMediator implements
 	        // Pre fetching invoke mediator flow
 	        String prefetchInvokeKey = key.evaluateValue(synCtx);
 	        InvokeMediator prefetchInvoke = (InvokeMediator) synCtx.getDefaultConfiguration(prefetchInvokeKey);
-	        RuntimeStatisticCollector.openLogForContinuation(synCtx, prefetchInvoke.getMediatorName());
+	        MediatorStatisticCollector.openLogForContinuation(synCtx, prefetchInvoke.getMediatorName());
 
 	        ContinuationState childContinuationState = continuationState.getChildContState();
 	        result = prefetchInvoke.mediate(synCtx, childContinuationState);
