+++ modules/core/src/main/java/org/apache/synapse/rest/RESTRequestHandler.java
@@ -22,8 +22,7 @@ import org.apache.axis2.Constants;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.synapse.MessageContext;
-import org.apache.synapse.aspects.ComponentType;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.APIStatisticCollector;
 import org.apache.synapse.core.axis2.Axis2MessageContext;
 import org.apache.synapse.rest.version.DefaultStrategy;
 
@@ -112,7 +111,7 @@ public class RESTRequestHandler {
 
 	private void apiProcess(MessageContext synCtx, API api) {
 		if (!synCtx.isResponse()) {
-			RuntimeStatisticCollector
+            APIStatisticCollector
 					.reportApiStatistics(synCtx, api.getAPIName(), api.getAspectConfiguration());
 		}
 		api.process(synCtx);
