+++ modules/core/src/main/java/org/apache/synapse/rest/Resource.java
@@ -26,7 +26,7 @@ import org.apache.synapse.ManagedLifecycle;
 import org.apache.synapse.Mediator;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.SynapseException;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.ResourceStatisticCollector;
 import org.apache.synapse.core.SynapseEnvironment;
 import org.apache.synapse.core.axis2.Axis2MessageContext;
 import org.apache.synapse.core.axis2.Axis2Sender;
@@ -263,7 +263,7 @@ public class Resource extends AbstractRESTProcessor implements ManagedLifecycle 
             if (getDispatcherHelper() != null) {
                 synCtx.setProperty(RESTConstants.REST_URL_PATTERN, getDispatcherHelper().getString());
             }
-	        RuntimeStatisticCollector.reportStatisticForResource(synCtx, name, null, true);
+            ResourceStatisticCollector.reportStatisticForResource(synCtx, name, null, true);
         }
 
         if (log.isDebugEnabled()) {
@@ -274,7 +274,7 @@ public class Resource extends AbstractRESTProcessor implements ManagedLifecycle 
         if (!synCtx.isResponse()) {
             String method = (String) synCtx.getProperty(RESTConstants.REST_METHOD);
             if (RESTConstants.METHOD_OPTIONS.equals(method) && sendOptions(synCtx)) {
-	            RuntimeStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
+                ResourceStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
 	            return;
             }
 
@@ -305,7 +305,7 @@ public class Resource extends AbstractRESTProcessor implements ManagedLifecycle 
         if (sequence != null) {
             registerFaultHandler(synCtx);
             sequence.mediate(synCtx);
-	        RuntimeStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
+            ResourceStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
             return;
         }
 
@@ -319,7 +319,7 @@ public class Resource extends AbstractRESTProcessor implements ManagedLifecycle 
                 throw new SynapseException("Specified sequence: " + sequenceKey + " cannot " +
                         "be found");
             }
-	        RuntimeStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
+            ResourceStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
             return;
         }
 
@@ -334,7 +334,7 @@ public class Resource extends AbstractRESTProcessor implements ManagedLifecycle 
         } else if (log.isDebugEnabled()) {
             log.debug("No in-sequence configured. Dropping the request.");
         }
-	    RuntimeStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
+        ResourceStatisticCollector.reportStatisticForResource(synCtx, name, null, false);
     }
 
     public void registerFaultHandler(MessageContext synCtx) {
