+++ modules/core/src/main/java/org/apache/synapse/util/MessageHelper.java
@@ -35,8 +35,8 @@ import org.apache.synapse.FaultHandler;
 import org.apache.synapse.MessageContext;
 import org.apache.synapse.SynapseConstants;
 import org.apache.synapse.SynapseException;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
-import org.apache.synapse.aspects.flow.statistics.util.StatisticMessageCountHolder;
+import org.apache.synapse.aspects.flow.statistics.collectors.AggregateMediatorStatisticCollector;
+import org.apache.synapse.aspects.flow.statistics.collectors.SplittingMediatorsStatisticCollector;
 import org.apache.synapse.aspects.statistics.ErrorLog;
 import org.apache.synapse.aspects.statistics.StatisticsLog;
 import org.apache.synapse.aspects.statistics.StatisticsRecord;
@@ -191,7 +191,7 @@ public class MessageHelper {
             }
         }
         newCtx.setMessageFlowTracingState(synCtx.getMessageFlowTracingState());
-        RuntimeStatisticCollector.setCloneProperties(synCtx, newCtx);
+        SplittingMediatorsStatisticCollector.setCloneProperties(synCtx, newCtx);
         return newCtx;
     }
 
@@ -313,7 +313,7 @@ public class MessageHelper {
         }
 
         newCtx.setMessageFlowTracingState(synCtx.getMessageFlowTracingState());
-        RuntimeStatisticCollector.setAggregateProperties(synCtx, newCtx);
+        AggregateMediatorStatisticCollector.setAggregateProperties(synCtx, newCtx);
 
         return newCtx;
     }
